pyproject.toml
[project]
name = "qr-based-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "gunicorn>=23.0.0",
    "mysql-connector-python>=9.4.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "python-dotenv>=1.1.1",
]
